Rename package apis/v1alpha1 to apis/v1
Since the API is already served under the "/api/v1" route, there is
little reason to version the internal types as "v1alpha1". The
versioning should probably be consistent.
