+++ apis/v1/auth.go
@@ -12,19 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v1alpha1
+package v1
 
-type FrontendAuthSettings struct {
-	BasicEnabled     bool   `json:"basicEnabled"`
-	OIDCEnabled      bool   `json:"oidcEnabled"`
-	OIDCProviderName string `json:"oidcProviderName,omitempty"`
-}
-
-// FrontendSettings are global settings exposed to the frontend, which can be
-// used to render some pages appropriately. These settings are not user-specific
-// and not confidential (the API for these settings is not protected by any auth
-// mechanism).
-type FrontendSettings struct {
-	Version string               `json:"version"`
-	Auth    FrontendAuthSettings `json:"auth"`
+type Token struct {
+	TokenType   string `json:"tokenType"`
+	AccessToken string `json:"accessToken"`
+	ExpiresIn   int64  `json:"expiresIn"`
 }
