+++ apis/v1/password.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v1alpha1
+package v1
 
 type UpdatePassword struct {
 	// base64-encoded passwords
