+++ test/e2e/data.go
@@ -18,7 +18,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 
-	apisv1alpha1 "antrea.io/antrea-ui/apis/v1alpha1"
+	apisv1 "antrea.io/antrea-ui/apis/v1"
 )
 
 const (
@@ -33,5 +33,5 @@ var (
 	// the host address for accessing the Antrea UI
 	host string
 
-	settings *apisv1alpha1.FrontendSettings
+	settings *apisv1.FrontendSettings
 )
